requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.27.2
pydantic==2.10.4
pydantic-settings==2.7.0
azure-cognitiveservices-speech==1.41.1
